create util-git libraries
