+++ libs/domains/organizations/feature/src/lib/git-repository-setting/git-repository-setting.tsx
@@ -3,7 +3,7 @@ import { Controller, useFormContext } from 'react-hook-form'
 import { useParams } from 'react-router-dom'
 import { InputSelect, LoaderSpinner } from '@qovery/shared/ui'
 import { upperCaseFirstLetter } from '@qovery/shared/util-js'
-import { getGitTokenValue } from '../git-provider-setting/git-provider-setting'
+import { getGitTokenValue } from '@qovery/util-git'
 import { useRepositories } from '../hooks/use-repositories/use-repositories'
 
 export interface GitRepositorySettingProps {
