+++ libs/pages/services/src/lib/feature/page-job-create-feature/step-summary-feature/step-summary-feature.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from 'react'
 import { useDispatch } from 'react-redux'
 import { useNavigate, useParams } from 'react-router-dom'
 import { createApplication, postApplicationActionsDeploy } from '@qovery/domains/application'
-import { getGitTokenValue, useContainerRegistry } from '@qovery/domains/organizations/feature'
+import { useContainerRegistry } from '@qovery/domains/organizations/feature'
 import { useImportVariables } from '@qovery/domains/variables/feature'
 import { type JobType, ServiceTypeEnum } from '@qovery/shared/enums'
 import {
@@ -26,6 +26,7 @@ import { FunnelFlowBody } from '@qovery/shared/ui'
 import { useDocumentTitle } from '@qovery/shared/util-hooks'
 import { buildGitRepoUrl } from '@qovery/shared/util-js'
 import { type AppDispatch } from '@qovery/state/store'
+import { getGitTokenValue } from '@qovery/util-git'
 import StepSummary from '../../../ui/page-job-create/step-summary/step-summary'
 import { useJobContainerCreateContext } from '../page-job-create-feature'
