+++ libs/shared/util-git/README.md
@@ -1,0 +1,7 @@
+# util-git
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test util-js` to execute the unit tests via [Jest](https://jestjs.io).
