+++ libs/shared/util-git/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/util-git'
