+++ tsconfig.base.json
@@ -75,7 +75,8 @@
       "@qovery/shared/util-web-sockets": ["libs/shared/util-web-sockets/src/index.ts"],
       "@qovery/shared/utils": ["libs/shared/utils/src/index.ts"],
       "@qovery/state/store": ["libs/state/store/src/index.ts"],
-      "@qovery/state/util-queries": ["libs/state/util-queries/src/index.ts"]
+      "@qovery/state/util-queries": ["libs/state/util-queries/src/index.ts"],
+      "@qovery/util-git": ["libs/shared/util-git/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
